fix: move the RepositoryConfig types to the APIs package
