+++ pkg/apis/lighthouse/v1alpha1/repoconfig_config_types.go
