+++ pkg/apis/lighthouse/v1alpha1/repoconfig_plugins_types.go
