+++ pkg/apis/lighthouse/v1alpha1/repoconfig_types.go
